pyproject.toml
[project]
name = "gemini-playground"
version = "0.1.0"
description = "A tool for generating image backgrounds using xAI and Gemini APIs."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai==1.5.0",
    "google-generativeai==0.8.4",
    "gradio==5.20.1",
    "openai==1.66.3",
]
